Algos solutions
