+++ src/algoexpertProblems/MinimumWaitingTime.java
@@ -1,0 +1,5 @@
+package algoexpertProblems;
+
+public class MinimumWaitingTime {
+
+}
